HKISD-226: refactor sapSlices to one file
