+++ src/App.tsx
@@ -27,7 +27,7 @@ import { CustomContextMenu } from './components/CustomContextMenu';
 import { getCoordinationGroupsThunk, getForcedToFrameGroupsThunk, getPlanningGroupsThunk } from './reducers/groupSlice';
 import { getHashTagsThunk } from './reducers/hashTagsSlice';
 import { clearLoading, setLoading } from './reducers/loaderSlice';
-import { getSapCostsThunk } from './reducers/sapCostSlice';
+import { getSapCostsThunk,getSapCurrentYearThunk } from './reducers/sapCostSlice';
 import moment from 'moment';
 import 'moment/locale/fi';
 import ScrollHandler from './components/shared/ScrollHandler';
@@ -47,7 +47,6 @@ import MaintenanceView from './views/Maintenance';
 import { AppDispatch } from './store';
 import AdminForcedToFrame from './components/Admin/AdminForcedToFrame/AdminForcedToFrame';
 import { getAppStateValuesThunk } from './reducers/appStateValueSlice';
-import { getSapCurrentYearThunk } from './reducers/sapCurrentYearSlice';
 
 const LOADING_APP_ID = 'loading-app-data';
